Added lateral controls and augmented camera tracking
Added lateral controls to the plane, enabling it to steer from side to
side using the arrow keys. This is only permitted during the “Game”
state (not while taking off, idle, or crashed). Also improved the
camera’s tracking abilities; It now has a blend tree to define the
position, and projection size according to it’s current take-off state,
as well as a broadened view for main game.